refactor: streamline TradingGuardWrapper and TradingPanel components to use useShowDepositWithdrawModal hook
- Replaced direct dialog handling with useShowDepositWithdrawModal for improved code clarity and reusability.
- Removed unused imports and optimized component structure for better maintainability.
- Enhanced enableTrading logic to utilize the new modal hook, ensuring consistent deposit handling across components.
