+++ packages/kit/src/views/Perp/components/TradingGuardWrapper.tsx
@@ -1,22 +1,17 @@
 import type { ReactNode } from 'react';
 import { memo, useCallback, useMemo } from 'react';
 
-import { noop } from 'lodash';
-
-import { Button, SizableText, useInTabDialog } from '@onekeyhq/components';
+import { Button, SizableText } from '@onekeyhq/components';
 import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
 import {
   usePerpsAccountLoadingInfoAtom,
   usePerpsActiveAccountAtom,
   usePerpsActiveAccountIsAgentReadyAtom,
-  usePerpsActiveAccountStatusAtom,
 } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 import { appLocale } from '@onekeyhq/shared/src/locale/appLocale';
 
-import { useHyperliquidActions } from '../../../states/jotai/contexts/hyperliquid';
-
-import { showDepositWithdrawDialog } from './TradingPanel/modals/DepositWithdrawModal';
+import { useShowDepositWithdrawModal } from '../hooks/useShowDepositWithdrawModal';
 
 interface ITradingGuardWrapperProps {
   children?: ReactNode;
@@ -32,7 +27,6 @@ function TradingGuardWrapperInternal({
   const [perpsAccount] = usePerpsActiveAccountAtom();
   const [perpsAccountLoading] = usePerpsAccountLoadingInfoAtom();
   const [{ isAgentReady }] = usePerpsActiveAccountIsAgentReadyAtom();
-
   // Memoize account info to optimize callback dependencies
   const accountInfo = useMemo(
     () => ({
@@ -41,7 +35,7 @@ function TradingGuardWrapperInternal({
     }),
     [perpsAccount.accountAddress, perpsAccount.accountId],
   );
-  const dialogInTab = useInTabDialog();
+  const { showDepositWithdrawModal } = useShowDepositWithdrawModal();
   const enableTrading = useCallback(async () => {
     try {
       const status =
@@ -51,17 +45,16 @@ function TradingGuardWrapperInternal({
         accountInfo.accountAddress &&
         accountInfo.accountId
       ) {
-        void showDepositWithdrawDialog(
-          {
-            actionType: 'deposit',
-          },
-          dialogInTab,
-        );
+        await showDepositWithdrawModal('deposit');
       }
     } catch (error) {
       console.error('[TradingGuardWrapper] Enable trading failed:', error);
     }
-  }, [accountInfo.accountAddress, accountInfo.accountId, dialogInTab]);
+  }, [
+    accountInfo.accountAddress,
+    accountInfo.accountId,
+    showDepositWithdrawModal,
+  ]);
 
   const shouldShowEnableTrading = useMemo(() => {
     return forceShowEnableTrading || isAgentReady === false;
