+++ packages/kit/src/views/Perp/components/TradingPanel/components/PerpsHeaderRight.tsx
@@ -8,12 +8,10 @@ import {
   IconButton,
   SizableText,
   XStack,
-  useInTabDialog,
   useMedia,
 } from '@onekeyhq/components';
 import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
 import { WalletConnectionForWeb } from '@onekeyhq/kit/src/components/TabPageHeader/components/WalletConnectionGroup';
-import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
 import {
   usePerpsActiveOpenOrdersAtom,
   usePerpsActivePositionAtom,
@@ -31,15 +29,12 @@ import {
   usePerpsActiveOrderBookOptionsAtom,
 } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
-import platformEnv from '@onekeyhq/shared/src/platformEnv';
-import { EModalRoutes } from '@onekeyhq/shared/src/routes';
-import { EModalPerpRoutes } from '@onekeyhq/shared/src/routes/perp';
 import { ETabRoutes } from '@onekeyhq/shared/src/routes/tab';
 
 import { usePerpsAssetCtx } from '../../../hooks/usePerpsAssetCtx';
 import { usePerpsMidPrice } from '../../../hooks/usePerpsMidPrice';
+import { useShowDepositWithdrawModal } from '../../../hooks/useShowDepositWithdrawModal';
 import { PerpSettingsButton } from '../../PerpSettingsButton';
-import { showDepositWithdrawDialog } from '../modals/DepositWithdrawModal';
 
 import { PerpsAccountNumberValue } from './PerpsAccountNumberValue';
 
@@ -97,27 +92,17 @@ function DebugButton() {
 function DepositButton() {
   const { gtSm } = useMedia();
   const [accountSummary] = usePerpsActiveAccountSummaryAtom();
-  const navigation = useAppNavigation();
   const accountValue = accountSummary?.accountValue;
   const intl = useIntl();
   const [activeAccount] = usePerpsActiveAccountAtom();
-  const dialogInTab = useInTabDialog();
+  const { showDepositWithdrawModal } = useShowDepositWithdrawModal();
   const content = activeAccount.accountAddress ? (
     <Badge
       borderRadius="$full"
       size="medium"
       variant="secondary"
       onPress={async () => {
-        if (gtSm) {
-          await showDepositWithdrawDialog(
-            { actionType: 'deposit' },
-            dialogInTab,
-          );
-        } else {
-          navigation.pushModal(EModalRoutes.PerpModal, {
-            screen: EModalPerpRoutes.MobileDepositWithdrawModal,
-          });
-        }
+        await showDepositWithdrawModal('deposit');
       }}
       alignItems="center"
       justifyContent="center"
